+++ Page Navigation App/View/Orders.xaml.cs
@@ -67,35 +67,41 @@ namespace Page_Navigation_App.View
         
         void EditOrder(object sender, ExecutedRoutedEventArgs e)
         {
-
-            var (list,err) = RW_Order.ReadwithID(e.Parameter.ToString(), Paths.sqlite_path);
-            if (list.Count==1)
+            if (Userhandling.GrantPermission(1, true))
             {
-                Edit_Order editOrder = new Edit_Order(list[0].ID,list[0].Customer,list[0].Description,list[0].RessourcesID,list[0].EndDate);
+                var (list, err) = RW_Order.ReadwithID(e.Parameter.ToString(), Paths.sqlite_path);
+                if (list.Count == 1)
+                {
+                    Edit_Order editOrder = new Edit_Order(list[0].ID, list[0].Customer, list[0].Description,
+                        list[0].RessourcesID, list[0].EndDate);
+                    editOrder.Owner = Application.Current.MainWindow;
+                    editOrder.ShowDialog();
+                    Load_Data(true);
+                }
+            }
+        }
+        
+        void AddOrder(object sender, RoutedEventArgs e)
+        {
+            if (Userhandling.GrantPermission(1, true))
+            {
+                Edit_Order editOrder = new Edit_Order("", "Customer", "Description", "Ressources", "Date");
                 editOrder.Owner = Application.Current.MainWindow;
                 editOrder.ShowDialog();
                 Load_Data(true);
             }
         }
         
-        void AddOrder(object sender, RoutedEventArgs e)
-        {
-            
-            Edit_Order editOrder = new Edit_Order("","Customer","Description","Ressources","Date");
-            editOrder.Owner = Application.Current.MainWindow;
-            editOrder.ShowDialog();
-            Load_Data(true);
-
-        }
-        
         void DeleteOrder(object sender, ExecutedRoutedEventArgs e)
         {
-            //hier auch noch Kundennamen mitgeben
-            Delete_Order deleteOrder = new Delete_Order(e.Parameter.ToString());
-            deleteOrder.Owner = Application.Current.MainWindow;
-            deleteOrder.ShowDialog();
-            Load_Data(true);
-
+            if (Userhandling.GrantPermission(1, true))
+            {
+                //hier auch noch Kundennamen mitgeben
+                Delete_Order deleteOrder = new Delete_Order(e.Parameter.ToString());
+                deleteOrder.Owner = Application.Current.MainWindow;
+                deleteOrder.ShowDialog();
+                Load_Data(true);
+            }
         }
 
         private void TextBoxFilter_OnTextChanged(object sender, TextChangedEventArgs e)
